+++ frontend/src/pages/analytics/metadata/events/MetadataEventSplitPanel.tsx
@@ -17,11 +17,9 @@ import {
   Button,
   ColumnLayout,
   FormField,
-  Pagination,
   SpaceBetween,
   SplitPanel,
   StatusIndicator,
-  Table,
   Tabs,
   TextContent,
   Textarea,
